app/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
cffi==1.14.6
chardet==4.0.0
click==8.0.1
discord-py-interactions==3.0.2
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pycparser==2.20
PyNaCl==1.4.0
pytube==11.0.1
regex==2021.8.28
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
yarl==1.6.3
youtube-dl==2021.6.6
